inital docker-compose yamls upload
includes a yaml for locally-built images (default, docker-compose.yml), images in a public repository (public,docker-compose.yml), and images in a private repository (private.docker-compose.yml). The private repository is almost assuredly inaccessible today, but provides an example of how it would look if you used an ECR (AWS) repository.